CompletionLineEdit in export/import
it's in options.py now, from where it can be imported
in the ui files it uses CompletionLineEdit directly now
