+++ projectgenerator/gui/import_data.py
@@ -21,7 +21,7 @@
 import webbrowser
 
 from projectgenerator.gui.ili2db_options import Ili2dbOptionsDialog
-from projectgenerator.gui.options import OptionsDialog
+from projectgenerator.gui.options import OptionsDialog, CompletionLineEdit
 from projectgenerator.gui.multiple_models import MultipleModelsDialog
 from projectgenerator.libili2db.iliimporter import JavaNotFoundError
 from projectgenerator.libili2db.ilicache import IliCache, ModelCompleterDelegate
@@ -111,6 +111,8 @@ class ImportDataDialog(QDialog, DIALOG_UI):
         self.gpkg_file_line_edit.setValidator(gpkgFileValidator)
 
         self.ili_models_line_edit.setPlaceholderText(self.tr('[Search model in repository]'))
+        self.ili_models_line_edit.textChanged.connect(self.complete_models_completer)
+        self.ili_models_line_edit.punched.connect(self.complete_models_completer)
         self.pg_host_line_edit.textChanged.connect(
             self.validators.validate_line_edits)
         self.pg_host_line_edit.textChanged.emit(self.pg_host_line_edit.text())
@@ -350,6 +352,13 @@ class ImportDataDialog(QDialog, DIALOG_UI):
                 self.base_configuration.save(settings)
         else:
             QDesktopServices.openUrl(link)
+
+    def complete_models_completer(self):
+        if not self.ili_models_line_edit.text():
+            self.ili_models_line_edit.completer().setCompletionMode(QCompleter.UnfilteredPopupCompletion)
+            self.ili_models_line_edit.completer().complete()
+        else:
+            self.ili_models_line_edit.completer().setCompletionMode(QCompleter.PopupCompletion)
 
     def update_models_completer(self):
         completer = QCompleter(self.ilicache.model, self.ili_models_line_edit)
