+++ projectgenerator/ui/export.ui
@@ -88,7 +88,7 @@
            </spacer>
           </item>
           <item row="1" column="2" colspan="2">
-           <widget class="QLineEdit" name="ili_models_line_edit"/>
+           <widget class="CompletionLineEdit" name="ili_models_line_edit"/>
           </item>
           <item row="0" column="0">
            <widget class="QLabel" name="label">
@@ -301,6 +301,13 @@
    </item>
   </layout>
  </widget>
+ <customwidgets>
+  <customwidget>
+   <class>CompletionLineEdit</class>
+   <header>projectgenerator.gui.options</header>
+   <container>1</container>
+  </customwidget>
+ </customwidgets>
  <tabstops>
   <tabstop>type_combo_box</tabstop>
   <tabstop>xtf_file_line_edit</tabstop>
