+++ projectgenerator/ui/generate_project.ui
@@ -125,7 +125,7 @@
            </widget>
           </item>
           <item row="2" column="1" colspan="2">
-           <widget class="QLineEdit" name="ili_models_line_edit_dummy"/>
+           <widget class="CompletionLineEdit" name="ili_models_line_edit"/>
           </item>
           <item row="0" column="1" colspan="2">
            <widget class="QLineEdit" name="ili_file_line_edit">
@@ -358,12 +358,17 @@
    <header>qgis.gui</header>
    <container>1</container>
   </customwidget>
+  <customwidget>
+   <class>CompletionLineEdit</class>
+   <header>projectgenerator.gui.options</header>
+   <container>1</container>
+  </customwidget>
  </customwidgets>
  <tabstops>
   <tabstop>type_combo_box</tabstop>
   <tabstop>ili_file_line_edit</tabstop>
   <tabstop>ili_file_browse_button</tabstop>
-  <tabstop>ili_models_line_edit_dummy</tabstop>
+  <tabstop>ili_models_line_edit</tabstop>
   <tabstop>multiple_models_button</tabstop>
   <tabstop>crsSelector</tabstop>
   <tabstop>ili2db_options_button</tabstop>
